+++ api/reviews.js
@@ -11,38 +11,46 @@ const {
   insertNewReview,
   getReviewById,
   replaceReviewById,
-  deleteReviewById
+  deleteReviewById,
+  verifyReviewOwner
 } = require('../models/review');
+const { requireAuthentication } = require('../lib/auth');
 
 /*
  * Route to create a new review.
  */
-router.post('/', async (req, res) => {
+router.post('/', requireAuthentication, async (req, res) => {
   if (validateAgainstSchema(req.body, ReviewSchema)) {
-    try {
-      /*
-       * Make sure the user is not trying to review the same business twice.
-       * If they're not, then insert their review into the DB.
-       */
-      const alreadyReviewed = await hasUserReviewedBusiness(req.body.userid, req.body.businessid);
-      if (alreadyReviewed) {
-        res.status(403).send({
-          error: "User has already posted a review of this business"
-        });
-      } else {
-        const id = await insertNewReview(req.body);
-        res.status(201).send({
-          id: id,
-          links: {
-            review: `/reviews/${id}`,
-            business: `/businesses/${req.body.businessid}`
-          }
+    if (req.user != null && (req.user.admin === 1 || req.user.sub == req.body.userid)) {
+      try {
+        /*
+         * Make sure the user is not trying to review the same business twice.
+         * If they're not, then insert their review into the DB.
+         */
+        const alreadyReviewed = await hasUserReviewedBusiness(req.body.userid, req.body.businessid);
+        if (alreadyReviewed) {
+          res.status(403).send({
+            error: "User has already posted a review of this business"
+          });
+        } else {
+          const id = await insertNewReview(req.body);
+          res.status(201).send({
+            id: id,
+            links: {
+              review: `/reviews/${id}`,
+              business: `/businesses/${req.body.businessid}`
+            }
+          });
+        }
+      } catch (err) {
+        console.error(err);
+        res.status(500).send({
+          error: "Error inserting review into DB.  Please try again later."
         });
       }
-    } catch (err) {
-      console.error(err);
-      res.status(500).send({
-        error: "Error inserting review into DB.  Please try again later."
+    } else {
+      res.status(403).send({
+        error: "Unauthorized to perform the specified action"
       });
     }
   } else {
@@ -74,41 +82,47 @@ router.get('/:id', async (req, res, next) => {
 /*
  * Route to update a review.
  */
-router.put('/:id', async (req, res, next) => {
+router.put('/:id', requireAuthentication, async (req, res, next) => {
   if (validateAgainstSchema(req.body, ReviewSchema)) {
-    try {
-      /*
-       * Make sure the updated review has the same businessID and userID as
-       * the existing review.  If it doesn't, respond with a 403 error.  If the
-       * review doesn't already exist, respond with a 404 error.
-       */
-      const id = parseInt(req.params.id);
-      const existingReview = await getReviewById(id);
-      if (existingReview) {
-        if (req.body.businessid === existingReview.businessid && req.body.userid === existingReview.userid) {
-          const updateSuccessful = await replaceReviewById(id, req.body);
-          if (updateSuccessful) {
-            res.status(200).send({
-              links: {
-                business: `/businesses/${req.body.businessid}`,
-                review: `/reviews/${id}`
-              }
-            });
+    if (req.user != null && (req.user.admin === 1 || req.user.sub === req.body.userid)) {
+      try {
+        /*
+         * Make sure the updated review has the same businessID and userID as
+         * the existing review.  If it doesn't, respond with a 403 error.  If the
+         * review doesn't already exist, respond with a 404 error.
+         */
+        const id = parseInt(req.params.id);
+        const existingReview = await getReviewById(id);
+        if (existingReview) {
+          if (req.body.businessid === existingReview.businessid && req.body.userid === existingReview.userid) {
+            const updateSuccessful = await replaceReviewById(id, req.body);
+            if (updateSuccessful) {
+              res.status(200).send({
+                links: {
+                  business: `/businesses/${req.body.businessid}`,
+                  review: `/reviews/${id}`
+                }
+              });
+            } else {
+              next();
+            }
           } else {
-            next();
+            res.status(403).send({
+              error: "Updated review must have the same businessID and userID"
+            });
           }
         } else {
-          res.status(403).send({
-            error: "Updated review must have the same businessID and userID"
-          });
+          next();
         }
-      } else {
-        next();
+      } catch (err) {
+        console.error(err);
+        res.status(500).send({
+          error: "Unable to update review.  Please try again later."
+        });
       }
-    } catch (err) {
-      console.error(err);
-      res.status(500).send({
-        error: "Unable to update review.  Please try again later."
+    } else {
+      res.status(403).send({
+        error: "Unauthorized to perform the specified action"
       });
     }
   } else {
@@ -121,18 +135,26 @@ router.put('/:id', async (req, res, next) => {
 /*
  * Route to delete a review.
  */
-router.delete('/:id', async (req, res, next) => {
-  try {
-    const deleteSuccessful = await deleteReviewById(parseInt(req.params.id));
-    if (deleteSuccessful) {
-      res.status(204).end();
-    } else {
-      next();
+router.delete('/:id', requireAuthentication, async (req, res, next) => {
+  const id = parseInt(req.params.id);
+  const review = await verifyReviewOwner(id, req.user.sub);
+  if (req.user != null && (req.user.admin === 1 || review)) {
+    try {
+      const deleteSuccessful = await deleteReviewById(parseInt(req.params.id));
+      if (deleteSuccessful) {
+        res.status(204).end();
+      } else {
+        next();
+      }
+    } catch (err) {
+      console.error(err);
+      res.status(500).send({
+        error: "Unable to delete review.  Please try again later."
+      });
     }
-  } catch (err) {
-    console.error(err);
-    res.status(500).send({
-      error: "Unable to delete review.  Please try again later."
+  } else {
+    res.status(403).send({
+      error: "Unauthorized to perform the specified action"
     });
   }
 });
